Add Authentication Middleware and Controller Validation to Goal Router

+++ backend/controller/goals/goalController.js
@@ -1,4 +1,5 @@
 import Goal from "../../models/goalModel.js";
+import User from "../../models/userModel.js";
 
 // POST create a new goal
 export const addGoal = async (req, res) => {
@@ -11,6 +12,7 @@ export const addGoal = async (req, res) => {
 
     const newGoal = await Goal.create({
       text: req.body.text,
+      user: req.user.id,
     });
 
     res.status(201).json(newGoal);
@@ -22,7 +24,7 @@ export const addGoal = async (req, res) => {
 // GET all goals
 export const getGoals = async (req, res) => {
   try {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
 
     res.status(200).json(goals);
   } catch (error) {
@@ -36,10 +38,20 @@ export const getGoalById = async (req, res) => {
     const goal = await Goal.findById(req.params.id);
 
     if (!goal) {
-      res.status(404).json({ message: "Goal not found" });
+      return res.status(404).json({ message: "Goal not found" });
     }
 
-    res.status(200).json(goal);
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+      return res.status(401).json({ message: "User not found" });
+    }
+
+    if (goal.user.toString() !== user.id) {
+      return res.status(401).json({ message: "User not authorized" });
+    }
+
+    return res.status(200).json(goal);
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
@@ -52,6 +64,16 @@ export const updateGoal = async (req, res) => {
 
     if (!goalToUpdate) {
       return res.status(404).json({ message: "Goal not found" });
+    }
+
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+      return res.status(401).json({ message: "User not found" });
+    }
+
+    if (goal.user.toString() !== user.id) {
+      return res.status(401).json({ message: "User not authorized" });
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -69,10 +91,18 @@ export const deleteGoal = async (req, res) => {
   try {
     const goalToDelete = await Goal.findById(req.params.id);
 
-    console.log(goalToDelete);
-
     if (!goalToDelete) {
       res.status(400).json({ message: "Goal not found" });
+    }
+
+    const user = await User.findById(req.params.id);
+
+    if (!user) {
+      return res.json({ message: "User not found" });
+    }
+
+    if (goal.user.toString() !== user.id) {
+      res.status(401).json({ message: "User not authorized" });
     }
 
     await Goal.deleteOne(goalToDelete);
